SCRD-2040 Generate lowercase /etc/hosts contents
The 0.20.0 version of eventlet, which is shipped with Pike,
considers the /etc/hosts contents as case-sensitive and causes
various openstack services to fail with "[Errno -2] No address
found" errors while trying to resolve Ardana generated hostnames.
As a quick fix for this problem, the config-processor
generated hosts file has been updated to include a lowercase
version of every hostname.

Change-Id: Ib15e2334ee08cf45955770b65d005928eb44e126