SCPL-339 make source match patch in IBS
The IBS build add the line:
global_vars['deployer_media_legacy_layout'] = False

to:
AnsibleAllVarsBuilder.py

The souce code has:
global_vars['deployer_media_legacy_layout'] = True

The source change was made after the patch in IBS was written, so tha 
patch can no longer be applied, I think the code in master
was meant to suuport the pre-ardana media layout, but with the
Cloud8 test, we are using the new lay put.

This change passes on it own, because we don't run deploy tests
on a-c-p, but I have included this review
as a depens on in review which builds the a-c-p rmp and aand installs
it and tests it. And with this change the current a-c-p  code in
master passes all tests.



Change-Id: I8875bdf35a4700fc0847432dc21682baf5af411a