SCRD-5520 Workaround for python-future html issue
The ardana-configuration-processor is implemented to use the python
module provided by the python-html package, which is a single file
python module, html.py, found under the site-packages area.

However once the python-future package is installed, we see problems
runningthe config processor because the python-future module installs
a directory based html package/module hierarchy under site-pacages,
which the Python import handling prefers to the single file html.py
module from the python-html package.

To workaround this we catch the import error and dynamically load
the htlm.py based version of the module explicitly, after locating
it in the system module search path.

Change-Id: I349cffa7706676438a7a8e8de16f9c989e7c31b7