Warn on unsupported physnet mappings (SCRD-7979)
Relying on a neutron provider network bridge to provide management
network access is unsupported and can result in management network
connectivity being interrupted. This change emits warnings when
using the "provider-physical-network" tag in a network group.
Because we can't clearly identify what the creator of the input
model intends to name the management network, the user is warned
when any network group is configured with the
"provider-physical-network" tag. This ensures the user is alerted to
potential issues with input models that may interrupt their control
plane connectivity.

Change-Id: I6edb5f5957074bf8572741fa24808afc203e263e